third_party/blink/web_tests/fast/layers/clip-rects-transformed-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
Test 1/2 for rdar://problem/7398987 Assertion failure in Layer::updateClipRects when a plugin’s enclosing layer is enclosed by a transformed layer.

This test should not cause an assertion faliure.
